% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_integral_on_tp.R
\name{compute_integral_on_tp}
\alias{compute_integral_on_tp}
\title{Compute integrals of TP}
\usage{
compute_integral_on_tp(trans_prob)
}
\arguments{
\item{trans_prob}{A list of transition probabilities.}
}
\value{
trans_prob, with int_p01_0 and int_p11 added.
}
\description{
Compute integrals of TP
}
