% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tp.R
\name{compute_tp}
\alias{compute_tp}
\title{Compute transition probability}
\usage{
compute_tp(survdat, t_gap = 1/12, method = "AJ", end_time, t)
}
\arguments{
\item{survdat}{A survTP object}
}
\value{
A list \itemize{
 \item p00_0, p01_0: \code{p00_0[i] = p00_0(0, t[i]), p01_0[i] = p01(0, t[i])}
 \item p01 and p11: Markov transition probabilites.
 \item time: time grid for searching the optimal schedule.
 \item int_p01_0, int_p11: integral of p01(0, s)ds and p11(t[i], s)ds
}
}
\description{
Compute transition probability
}
