% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_high_order.R
\name{dp_high_order}
\alias{dp_high_order}
\title{Dynamic Programming with J = 1 or 2 (under Markov assumption). dy_proc + find_schedule.}
\usage{
dp_high_order(tp_est, maxK, beta = 1, J = 1, echo = TRUE,
  only_maxK = TRUE)
}
\value{
\itemize{
 \item idx: indeces of the optimal solution
 \item val: solution value
 \item loss_est: estimated expected loss (with corresponding J)
}
}
\description{
Dynamic Programming with J = 1 or 2 (under Markov assumption). dy_proc + find_schedule.
}
