% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_proc.R
\name{dp_proc}
\alias{dp_proc}
\title{Dynamic Programming procedure
Algorithm 1 in the paper}
\usage{
dp_proc(tp_est, maxK, output_eta = FALSE)
}
\arguments{
\item{tp_est}{List of transition probability estimation.}

\item{maxK}{Integer.}

\item{output_eta}{Logical.}
}
\value{
List of \itemize{
 \item time, loss_const_term
 \item opt: C matrix in the paper
 \item opt_idx: argmin_{l} { C[k-1, t[l]] + eta[t[l], t[i]] }
 \item const_term: the constant term in the expected loss
 \item eta_0, eta (optional): eta function as defined in the paper, which is the partial loss.
}
}
\description{
Dynamic Programming procedure
Algorithm 1 in the paper
}
