% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_loss.R
\name{eval_loss}
\alias{eval_loss}
\title{Evaluating the (empirical) loss for a given schedule (in idx)}
\usage{
eval_loss(sched_idx, tp_est, beta = 1, J = Inf)
}
\value{
loss_est: loss evaluated result.
}
\description{
Evaluating the (empirical) loss for a given schedule (in idx)
}
