% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_schedule.R
\name{find_schedule}
\alias{find_schedule}
\title{Restore the optical screening time from the DP result}
\usage{
find_schedule(dp_res, K)
}
\value{
\itemize{
 \item idx: indeces of the optimal solution
 \item val: solution value
}
}
\description{
Restore the optical screening time from the DP result
}
